SpreadsheetFindDialogComponentSpreadsheetDataTableComponentCellRenderer extracted
